Refactored the code

+++ src/main/java/com/bookout/database/daointerfaces/BookingDAO.java
@@ -1,8 +1,10 @@
 package com.bookout.database.daointerfaces;
 
+
 import java.sql.SQLException;
 import java.util.List;
 
 public interface BookingDAO<T> extends ObjectInterface<T> {
-    List<T> getBookingsByRestaurantId(long restaurantId) throws SQLException;
+
+    List<T> getBookingsByRestaurantId(int restaurantId) throws SQLException;
 }
